bootenv: move unmarshalHashAlg to export_test.go
