+++ bootenv/export_test.go
@@ -26,6 +26,7 @@ import (
 	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/x509"
+	"encoding/asn1"
 	"errors"
 
 	"github.com/snapcore/secboot"
@@ -72,6 +73,43 @@ func MockLoadCurrenBootMode(f func() (string, error)) (restore func()) {
 
 func (d *KeyDataScope) TestSetVersion(version int) {
 	d.data.Version = version
+}
+
+func unmarshalHashAlg(s *cryptobyte.String) (hashAlg, error) {
+	var str cryptobyte.String
+
+	if !s.ReadASN1(&str, cryptobyte_asn1.SEQUENCE) {
+		return 0, errors.New("malformed input")
+	}
+
+	var oid asn1.ObjectIdentifier
+
+	if !str.ReadASN1ObjectIdentifier(&oid) {
+		return 0, errors.New("malformed Algorithm identifier")
+	}
+
+	var null uint8
+
+	if !str.ReadUint8(&null) {
+		return 0, errors.New("malformed input")
+	}
+
+	if len(oid) == len(sha1Oid) {
+		return hashAlg(crypto.SHA1), nil
+	}
+
+	switch oid[8] {
+	case sha224Oid[8]:
+		return hashAlg(crypto.SHA224), nil
+	case sha256Oid[8]:
+		return hashAlg(crypto.SHA256), nil
+	case sha384Oid[8]:
+		return hashAlg(crypto.SHA384), nil
+	case sha512Oid[8]:
+		return hashAlg(crypto.SHA512), nil
+	default:
+		return 0, errors.New("unsupported hash algorithm")
+	}
 }
 
 func UnmarshalAdditionalData(data []byte) (*additionalData, error) {
